+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [1.5.0](https://github.com/samtun/hackworld/compare/v1.4.0...v1.5.0) (2025-12-20)
+
+
+### Bug Fixes
+
+* use actual maxHp/maxTp values in XData UI instead of calculated values ([66d544e](https://github.com/samtun/hackworld/commit/66d544ec0293f18d657441cbca41922fc6af81d9))
+
+
+### Features
+
+* make level up heal player and increase HP and TP ([d18d19a](https://github.com/samtun/hackworld/commit/d18d19aa83000e7c50687ccf45564a38f22c0470))
+
+# [1.4.0](https://github.com/samtun/hackworld/compare/v1.3.0...v1.4.0) (2025-12-20)
+
+
+### Features
+
+* minor fixes ([ed24699](https://github.com/samtun/hackworld/commit/ed246991fc83cbf5e0fc828b228b06a77e30d4f5))
+* remove center obstacle from second dungeon ([b083be5](https://github.com/samtun/hackworld/commit/b083be5c27c7b52d566e0bd14d8d0e89f70f6d55))
+
+# [1.3.0](https://github.com/samtun/hackworld/compare/v1.2.0...v1.3.0) (2025-12-20)
+
+
+### Features
+
+* add controller navigation to death overlay and reload stage on retry ([d31afe4](https://github.com/samtun/hackworld/commit/d31afe47be1e7da80c25ace34711a8da9ac7541c))
+* implement player death system with respawn and lobby return ([b47e6ee](https://github.com/samtun/hackworld/commit/b47e6ee6ca27fbee0d424cdda4a508b81993a582))
+
 # [1.2.0](https://github.com/samtun/hackworld/compare/v1.1.0...v1.2.0) (2025-12-20)